rename some test files
