+++ tests/meson.build
@@ -8,7 +8,8 @@ test_sources = [
     'main_frame_test.cpp',
     'string_utils_test.cpp',
     'validator_test.cpp',
-    'vector_utils_test.cpp',
+    'string_test.cpp',
+    'vector_test.cpp',
 ]
 
 # build tests
